update cards hover states
