+++ src/pages/explore/wallets/_components/AdvanceUsageSection.tsx
@@ -11,7 +11,7 @@ export function AdvanceUsageSection() {
   const { t } = useTranslation(["page-downloads"]);
 
   return (
-    <Container className="mt-[120px] md:mt-[96px] lg:mt-[240px]">
+    <Container className="mt-[120px] md:mt-[96px] lg:mt-[240px] lg:mb-[240px] md:mb-[112px] mb-[72px]">
       <div className="flex flex-col gap-y-5">
         <div className="flex md:justify-center justify-start">
           <SectionTitle text="SUITABLE FOR ADVANCE USAGE" />
