+++ src/pages/explore/wallets/_components/icons/BraveIcon.tsx
