+++ src/pages/explore/wallets/_components/icons/FirefoxIcon.tsx
