+++ src/pages/explore/wallets/_components/icons/WindowsIcon.tsx
